Search Function Added

+++ src/model/InHouse.java
@@ -3,7 +3,7 @@ package model;
 public class InHouse extends Part{
     int machineId;
 
-    public InHouse(int id, String name, double stock, double price, int max, int min) {
+    public InHouse(int id, String name, int stock, double price, int max, int min) {
         super(id, name, stock, price, max, min);
     }
