+++ src/model/Outsourced.java
@@ -3,7 +3,7 @@ package model;
 public class Outsourced extends Part {
     String companyName;
 
-    public Outsourced(int id, String name, double stock, double price, int max, int min) {
+    public Outsourced(int id, String name, int stock, double price, int max, int min) {
         super(id, name, stock, price, max, min);
     }
 }
